refactor(js): merge entry.ts + worker.ts into a single runner.ts
The two worker scripts shared most of their logic (WASI sandbox setup,
wasm compile/instantiate, `thread-spawn` host import, printErr hook)
and only diverged on what they did after instantiation: the entry
worker called `wasm_main` and posted the report; thread workers waited
for `{startArg, tid, memory}` and called `wasi_thread_start`.

Collapse them into one `runner.ts` whose mode is selected by
`workerData.mode` ('entry' | 'thread'), and consolidate the duplicated
`worker.on('message')` dispatch in `index.ts` into a single
`attachCommonHandlers(w)` helper. Net -308 lines across `src/*.ts`
(2040 → 1732) with no behavior change — all 51 existing library + CLI
tests pass.
