+++ tsdown.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsdown';
 
 export default defineConfig({
-  entry: ['./src/index.ts', './src/cli.ts', './src/worker.ts', './src/entry.ts'],
+  entry: ['./src/index.ts', './src/cli.ts', './src/runner.ts'],
   format: ['esm', 'cjs'],
   dts: true,
   outDir: 'dist',
